aat Home

+++ MyTrade/ResultChart.cs
@@ -67,7 +67,7 @@ namespace MyTrade
     public class Quote
     {
         public List<double> low { get; set; }
-        public List<int> volume { get; set; }
+        public List<double> volume { get; set; }
         public List<double> close { get; set; }
         public List<double> high { get; set; }
         public List<double> open { get; set; }
